+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 * Fixed server crash with the cauldron provider
 
 ### 1.1.0
-* ProjectE Compat: show block EMC and EMC stored in tile entities
+* ProjectE Compat: Show block EMC and EMC stored in tile entities
+* Thaumcraft 6 Compat: Show smelting progress and fuel burn time in Essentia Smelteries